requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
python-dotenv==1.0.0
psycopg2-binary==2.9.11
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
